modularize serialization code
move some serialize / deserialize code into the serialization module
to get around bootstrap issues
