+++ base/Enums.jl
@@ -91,7 +91,7 @@ macro enum(T,syms...)
         immutable $(esc(T)) <: Enum
             val::$enumT
             function $(esc(typename))(x::Integer)
-                $(membershiptest(:x, values)) || enum_argument_error($(Meta.quot(typename)), x)
+                $(membershiptest(:x, values)) || enum_argument_error($(Expr(:quote, typename)), x)
                 new(x)
             end
         end
@@ -100,7 +100,7 @@ macro enum(T,syms...)
         Base.length{E<:$(esc(typename))}(x::Type{E}) = $(length(vals))
         Base.next{E<:$(esc(typename))}(x::Type{E},s) = (E($values[s]),s+1)
         Base.done{E<:$(esc(typename))}(x::Type{E},s) = s > $(length(values))
-        Base.names{E<:$(esc(typename))}(x::Type{E}) = [$(map(x->Meta.quot(x[1]), vals)...)]
+        Base.names{E<:$(esc(typename))}(x::Type{E}) = [$(map(x->Expr(:quote, (x[1])), vals)...)]
         Base.isless{E<:$(esc(typename))}(x::E, y::E) = isless(x.val, y.val)
         function Base.print{E<:$(esc(typename))}(io::IO,x::E)
             for (sym, i) in $vals
